+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- extension: registerEntityManager() and prependEntityManagerConfig() now accept an optional list of
+  entity manager dependencies. MigrateCommand migrates dependencies before the entity managers depending
+  on them.
+
 ## v0.1.246
 
 - Allow null address attributes for Helpers/Tools::createAddressArray()